Add some new shooter and caster-oriented arts
Imoen: A shallow shortbow which gives DEC_MANA.
Sharnga: A powerful lategame longbow based on the item from DCSS.
Punk: An acidic sling with the same source as Sharnga.
Trebuchet: A high damage sling which decreases speed.
Plinker: A weakish sling with the Stun ego.
Avalanche: A powerful sling with no increased damage, but 6 pval
    worth of XTRA_SHOTS.
Spellshard: A shallow dagger with modest, but useful abilities for
    casters.
Fightin' Stick: An [M wizardstaff with good dice, blows, and |M.
Elfcoat: A deep mithril chainmail with low weight, and excellent
    resists, also boosts SPELL_CAP and gives some stat boosts.
Ogre Blade: A comical club with VORPAL ego, not a bad weapon
    but mostly for laughs.
Guillotine: A heavy vorpal shield, for use by a specific class.